requirements.txt
fastapi==0.128.8
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.11.0
PyYAML==6.0.3
uvicorn==0.39.0
pytest==8.4.2
# Optional for production BigQuery integration:
# google-cloud-bigquery==3.40.1
